+++ python/tensorstudio/nn/__init__.py
@@ -2,15 +2,44 @@
 
 from __future__ import annotations
 
-from .functional import binary_cross_entropy, l1_loss, mse_loss, relu, sigmoid, tanh
-from .losses import BCELoss, L1Loss, MSELoss
-from .modules import Dropout, Flatten, Linear, Module, Parameter, ReLU, Sequential, Sigmoid, Tanh
+from .functional import (
+    binary_cross_entropy,
+    binary_cross_entropy_with_logits,
+    huber_loss,
+    l1_loss,
+    leaky_relu,
+    linear,
+    mse_loss,
+    relu,
+    sigmoid,
+    softplus,
+    tanh,
+)
+from .losses import BCELoss, BCEWithLogitsLoss, HuberLoss, L1Loss, MSELoss
+from .modules import (
+    Dropout,
+    Flatten,
+    Identity,
+    LeakyReLU,
+    Linear,
+    Module,
+    Parameter,
+    ReLU,
+    Sequential,
+    Sigmoid,
+    Softplus,
+    Tanh,
+)
 
 __all__ = [
     "BCELoss",
+    "BCEWithLogitsLoss",
     "Dropout",
     "Flatten",
+    "HuberLoss",
+    "Identity",
     "L1Loss",
+    "LeakyReLU",
     "Linear",
     "MSELoss",
     "Module",
@@ -18,11 +47,17 @@ __all__ = [
     "ReLU",
     "Sequential",
     "Sigmoid",
+    "Softplus",
     "Tanh",
     "binary_cross_entropy",
+    "binary_cross_entropy_with_logits",
+    "huber_loss",
     "l1_loss",
+    "leaky_relu",
+    "linear",
     "mse_loss",
     "relu",
     "sigmoid",
+    "softplus",
     "tanh",
 ]